数据库存储short 有问题，因此修改为int 增加TimePatterDesc类 TimeBaseServiceImpl类型修改后，修改月份。取int字节的0，1 位 去除去调试的部分

+++ app/src/main/java/cn/com/aiton/domain/GbtTimeBase.java
@@ -16,7 +16,7 @@ public class GbtTimeBase {
 
 	private int week;
 
-	private int day;
+	private long day;
 
 	private int scheduleId;
 
@@ -60,11 +60,11 @@ public class GbtTimeBase {
         this.week = week;
     }
 
-    public int getDay() {
+    public long getDay() {
         return day;
     }
 
-    public void setDay(int day) {
+    public void setDay(long day) {
         this.day = day;
     }
